editComment

+++ backend/models/comment.js
@@ -48,6 +48,17 @@ class Comment {
             comment_id
         );
 
+        const checkIfExisting = await db.query(
+            `SELECT * FROM comments
+            WHERE user_id = (SELECT id FROM users WHERE username = $1)
+            AND id = $2`,
+            [user.username, comment_id]
+        );
+
+        if (checkIfExisting.rows.length !== 1) {
+            throw new BadRequestError("You don't have this comment");
+        }
+
         const results = await db.query(
             `DELETE FROM comments
             WHERE user_id = (SELECT id FROM users WHERE username = $1)
@@ -56,10 +67,30 @@ class Comment {
             [user.username, comment_id]
         );
         console.log(results.rows[0]);
-        if (results.rows[0] === undefined) {
-            throw new BadRequestError("Not your comment");
-        }
+
         return results.rows[0];
+    }
+
+    static async editComment(user, comment) {
+        if (!user) {
+            throw new UnauthorizedError(`No user logged in`);
+        }
+
+        console.log(comment.comment, comment.id);
+
+        const results = await db.query(
+            `UPDATE comments
+            SET comment = $1
+            WHERE user_id = (SELECT id FROM users WHERE username = $2)
+            AND id = $3
+            RETURNING comment, id
+            `,
+            [comment.comment, user.username, comment.id]
+        );
+        console.log("results.rows[0]: ", results.rows[0]);
+        return results.rows[0];
+
+        // return results.rows[0];
     }
 }
